new structure for flask server
