+++ profiles.json
@@ -6,19 +6,19 @@
         "nbr_ratio_lr": 0
     },
     "301727996392505346": {
-        "score": 1827,
+        "score": 1847,
         "nbr_ratio_r": 1,
         "nbr_ratio_sr": 0,
         "nbr_ratio_lr": 1
     },
     "398141917495558168": {
-        "score": 326,
+        "score": 332,
         "nbr_ratio_r": 0,
         "nbr_ratio_sr": 0,
         "nbr_ratio_lr": 0
     },
     "456149783342481408": {
-        "score": 1088,
+        "score": 1098,
         "nbr_ratio_r": 0,
         "nbr_ratio_sr": 0,
         "nbr_ratio_lr": 0
